refactor(fmt): Move impls next to types
